+++ packages/react-widgets/src/index.js
@@ -12,4 +12,4 @@ export {
   geocodeStreetPoint,
   getScatter
 } from './models';
-export { default as useSourceFilters } from './utils/useSourceFilters';
+export { default as useSourceFilters } from './hooks/useSourceFilters';
